% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_street_addresses}
\alias{extract_street_addresses}
\title{Extract street address components}
\usage{
extract_street_addresses(street_addresses, dump_path = NULL)
}
\arguments{
\item{street_addresses}{List of street addresses}

\item{dump_path}{Location for dumping list of failed extractions}
}
\description{
Takes a list of street addresses and returns a list of structured address
components.
}
\examples{
source_df \%>\%
  left_join(
    extract_street_addresses(.$raw_street_addresses, dump_path = "/dump"),
    by = "raw_street_addresses",
    relationship = "one-to-many") # Raw addresses are often describing many addresses

}
