% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{unabbreviate_street_types}
\alias{unabbreviate_street_types}
\title{Unabbreviate street types}
\usage{
unabbreviate_street_types(abbrv_street_types)
}
\arguments{
\item{abbrv_street_types}{List of abbreviated street types}
}
\description{
Takes a list of abbreviated street types and returns a list of full street
types. Mostly a wrapper to make Python import easier.
}
